pyproject.toml
[project]
name = "agent-platform"
version = "0.1.0"
description = "FastAPI backend for unified agent orchestration"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
keywords = ["fastapi", "agents", "rag"]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.0.8",
    "supabase>=2.6.0",
    "google-genai>=1.47.0",
    "python-multipart>=0.0.6",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "sentry-sdk[fastapi]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-anyio>=0.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "docs*", "static*"]

[tool.pytest.ini_options]
anyio_backends = ["asyncio"]
